Use new parameters from callback. Store the logger in a wrapper and call getLevel and log function on the wrapper, strangely calling log from c++ directly on logger with extra named parameter resulted to an unexplicable crash.

+++ cpp/src/pylogging.h
@@ -8,7 +8,7 @@ class CppToPythonLogger
   public:
     static CppToPythonLogger* get();
 
-    static void logFromJava(int level, int timestamp, char* loggerName, char* message);
+    static void logFromJava(int level, double timestamp, char* loggerName, char* message);
 
     void setLogger(py::object pPythonLogger);
